Move RELOAD/tls config connection refresh tests from test_operations -> test_ssl so they are only run with SSL tests in CI
