+++ test/ssl/test.ini
@@ -2,6 +2,7 @@
 p0 = port=6666 host=localhost dbname=p0 user=bouncer pool_size=2
 p1 = port=6666 host=localhost dbname=p1 user=bouncer
 p7a= port=6666 host=localhost dbname=p7
+pTxnPool = port=6666 host=127.0.0.1 dbname=p0 user=bouncer pool_mode=transaction
 
 [pgbouncer]
 logfile = test.log
