+++ test/test.ini
@@ -51,8 +51,6 @@ replication = port=6666 host=127.0.0.1 dbname=replication
 ; commented out except for auto-database tests
 ;* = port=6666 host=127.0.0.1
 
-pTxnPool = port=6666 host=127.0.0.1 dbname=p0 user=bouncer pool_mode=transaction
-
 [users]
 maxedout = max_user_connections=3
 maxedout2 = max_user_connections=2
